+++ src/ska_low_mccs_pasd/fndh/fndh_device.py
@@ -82,7 +82,7 @@ class MccsFNDH(MccsBaseDevice[FndhComponentManager]):
     FaultOnThresholdDifference: Final = device_property(
         doc="Put the device in DevState FAULT if firmware and Tango thresholds differ.",
         dtype=bool,
-        default_value=False,
+        default_value=True,
     )
 
     # ---------